requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-SocketIO==5.0.1
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
pydantic==1.8.1
python-engineio==4.0.1
python-socketio==5.1.0
python-vlc==3.0.12117
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.4
websocket-client==0.58.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.3.0
